+++ protoc_plugin/lib/protoc.dart
@@ -15,6 +15,7 @@ import 'src/gen/dart_options.pb.dart';
 import 'src/gen/google/api/client.pb.dart';
 import 'src/gen/google/protobuf/compiler/plugin.pb.dart';
 import 'src/gen/google/protobuf/descriptor.pb.dart';
+import 'src/gen/google/protobuf/dart_edition_defaults.pb.dart';
 import 'src/linker.dart';
 import 'src/options.dart';
 import 'src/output_config.dart';
@@ -34,3 +35,12 @@ part 'src/paths.dart';
 part 'src/protobuf_field.dart';
 part 'src/service_generator.dart';
 part 'src/well_known_types.dart';
+
+final FeatureSetDefaults pluginFeatureSetDefaults =
+    FeatureSetDefaults.fromBuffer(
+      base64Decode(ProtobufInternalDartEditionDefaults),
+    );
+
+const Edition pluginMinSupportedEdition = Edition.EDITION_PROTO2;
+
+const Edition pluginMaxSupportedEdition = Edition.EDITION_2024;